+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Next
 - Fix for escaping identifiers correctly (#340)
 - Support for S3-compatible storage (#343)
 - Adapters can now know which columns were requested (#345)
+- Python 3.11 officially supported (#334)
 
 Version 1.2.0 - 2023-02-17
 ==========================